Do not delete .git_release during a deployment.
Instead, mark that the deployment is in progress.
This allows for better UI, and prevents displaying an
error message in the web UI when the page is loaded during
the copy step.

Also normalize the .git_release file date format, and cleanup
related code.

This is a slightly backward-incompatible change, since it changes
the format of .git_release. Most apps should be robust enough
to handle this.